[feedback] 드론위치 계산 로직 분리 (useDronePosition, 훅)
