+++ src/components/map/MapView.tsx
@@ -14,6 +14,7 @@ import React, { useEffect, useState } from "react";
 import { getColorFromId } from "@/utils/getColorFromId";
 import { mapCalculator } from "@/utils/mapCalculator";
 import useOperationData from "@/hooks/useOperationData";
+import useDronePosition from "@/hooks/useDronePosition";
 
 interface MapViewProps {
   progress: number;
@@ -38,62 +39,20 @@ export default function MapView({
   const [dronePositions, setDronePositions] = useState<
     { flightId: string; position: [number, number]; direction: number }[]
   >([]);
+  const updatedPositions = useDronePosition(
+    progress,
+    allTimestamps,
+    operationTimestamps,
+    operationLatlngs,
+  );
 
   useEffect(() => {
     setOperationLatlngs(updatedLatlngs);
   }, [updatedLatlngs]);
 
   useEffect(() => {
-    if (!allTimestamps || allTimestamps.length < 2) return;
-
-    const allStartTime = allTimestamps[0];
-    const allEndTime = allTimestamps[allTimestamps.length - 1];
-    const totalDuration = allEndTime - allStartTime;
-    const currentTime = allStartTime + (totalDuration * progress) / 100;
-
-    // 운행별 마커 위치 계산
-    const updatedPositions = selectedOperationId.map((id) => {
-      const timestamps = operationTimestamps[id] || [];
-      const positions = operationLatlngs[id] || [];
-
-      const startTime = timestamps[0];
-      const endTime = timestamps[timestamps.length - 1];
-
-      // 운행 시작 전
-      if (currentTime < startTime) {
-        return { flightId: id, position: positions[0], direction: 0 };
-      }
-
-      // 운행 중
-      if (currentTime >= startTime && currentTime <= endTime) {
-        // 개별 진행률 계산
-        const operationProgress =
-          ((currentTime - startTime) / (endTime - startTime)) * 100;
-
-        const position = mapCalculator.calculateDronePosition(
-          operationProgress,
-          positions,
-          timestamps,
-        );
-        const direction = mapCalculator.calculateDirection(
-          operationProgress,
-          positions,
-          timestamps,
-        );
-
-        return { flightId: id, position, direction };
-      }
-
-      // 운행 종료 후
-      return {
-        flightId: id,
-        position: positions[positions.length - 1],
-        direction: 0,
-      };
-    });
-
     setDronePositions(updatedPositions);
-  }, [progress, allTimestamps, operationTimestamps, operationLatlngs]);
+  }, [updatedPositions]);
 
   return (
     <div className="relative z-0 h-full w-full">
